Added fake data generation to database cli commands. Separated this from the setup commands found in setup.py (Fixes #3).
