+++ app/models/__init__.py
@@ -2,3 +2,5 @@ from .resource import *
 from .resource_category import *
 from .user import *
 from .role import *
+
+
