+++ app/models/resource.py
@@ -1,5 +1,4 @@
-from ..db import db
-
+from app import db
 
 class Resource(db.Model):
     """
