+++ app/models/resource_category.py
@@ -1,4 +1,4 @@
-from ..database import db
+from app import db
 
 
 class ResourceCategory(db.Model):
