+++ app/models/role.py
@@ -1,4 +1,4 @@
-from ..database import db
+from app import db
 
 class Permission:
     ADMINISTER = 0x01
