+++ app/models/user.py
@@ -4,9 +4,7 @@ from flask_login import UserMixin, AnonymousUserMixin
 from werkzeug.security import check_password_hash, generate_password_hash
 from itsdangerous import TimedJSONWebSignatureSerializer as Serializer
 from itsdangerous import BadSignature, SignatureExpired
-
-from ..database import db
-
+from . import db
 
 class User(UserMixin, db.Model):
     __tablename__ = 'users'
